fix: move error details out of packages/ui into kilo-ui only
Remove all changes from packages/ui (upstream). ErrorDetails component,
CSS, and stories now live entirely in packages/kilo-ui. Screenshot test
stories show expanded error details for visual regression testing.
