+++ packages/kilo-ui/src/styles/index.css
@@ -18,6 +18,7 @@
 @import "../components/dialog.css";
 @import "../components/diff-changes.css";
 @import "../components/dropdown-menu.css";
+@import "../components/error-details.css";
 @import "../components/icon-button.css";
 @import "../components/inline-input.css";
 @import "../components/list.css";
