+++ packages/ui/src/components/session-turn.css
@@ -75,63 +75,6 @@
     overflow-y: auto;
   }
 
-  .error-card [data-component="collapsible"] {
-    margin-top: 8px;
-  }
-
-  .error-details-trigger {
-    display: inline-flex;
-    align-items: center;
-    gap: 4px;
-    font-size: var(--font-size-xs, 11px);
-    opacity: 0.85;
-    cursor: pointer;
-    background: none;
-    border: none;
-    color: inherit;
-    padding: 0;
-  }
-
-  .error-details-trigger:hover {
-    opacity: 1;
-  }
-
-  .error-details {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    font-size: var(--font-size-xs, 11px);
-    margin-top: 4px;
-  }
-
-  .error-detail-row {
-    display: flex;
-    gap: 8px;
-    align-items: baseline;
-  }
-
-  .error-detail-label {
-    font-weight: var(--font-weight-medium, 500);
-    white-space: nowrap;
-    min-width: 70px;
-    flex-shrink: 0;
-  }
-
-  .error-detail-value {
-    word-break: break-all;
-  }
-
-  .error-detail-pre {
-    margin: 0;
-    max-height: 120px;
-    overflow-y: auto;
-    font-size: var(--font-size-xs, 11px);
-    white-space: pre-wrap;
-    word-break: break-all;
-    flex: 1;
-    min-width: 0;
-  }
-
   [data-slot="session-turn-assistant-content"] {
     width: 100%;
     min-width: 0;
