+++ packages/ui/src/components/session-turn.tsx
@@ -14,7 +14,6 @@ import { Collapsible } from "./collapsible"
 import { DiffChanges } from "./diff-changes"
 import { Icon } from "./icon"
 import { TextShimmer } from "./text-shimmer"
-import { ErrorDetails, hasErrorDetails } from "./error-details"
 import { createAutoScroll } from "../hooks"
 import { useI18n } from "../context/i18n"
 
@@ -486,18 +485,7 @@ export function SessionTurn(
                 </Show>
                 <Show when={error()}>
                   <Card variant="error" class="error-card">
-                    <div>{errorText()}</div>
-                    <Show when={hasErrorDetails(error())}>
-                      <Collapsible variant="ghost">
-                        <Collapsible.Trigger class="error-details-trigger">
-                          <span>{i18n.t("error.details.show")}</span>
-                          <Collapsible.Arrow />
-                        </Collapsible.Trigger>
-                        <Collapsible.Content>
-                          <ErrorDetails error={error()!} />
-                        </Collapsible.Content>
-                      </Collapsible>
-                    </Show>
+                    {errorText()}
                   </Card>
                 </Show>
               </div>
