+++ packages/ui/src/i18n/zh.ts
@@ -120,12 +120,4 @@ export const dict = {
   "ui.question.multiHint": "可多选",
   "ui.question.singleHint": "选择一个答案",
   "ui.question.custom.placeholder": "输入你的答案...",
-
-  "error.details.show": "详细信息",
-  "error.details.type": "类型",
-  "error.details.statusCode": "状态码",
-  "error.details.provider": "提供商",
-  "error.details.retryable": "可重试",
-  "error.details.retries": "重试次数",
-  "error.details.responseBody": "响应",
 } satisfies Partial<Record<Keys, string>>
